Improve name validation function and test long name
The functions validating schema object names are misleading and placed
in different files. This led to incorrect use of them and failures on
creating tables with long names.

This PR reorganizes the name validation functions, so they are not
confused and misused. The error messages are fixed accordantly.

This PR adds a long table name test with non-alpha-numeric table name
to improve the test coverage.