Separate length and character validation for names
Since different types of names, e.g., keyspace, table, index, column,
are checked for different lengths, the length validation is removed
from isValidName function, which only checks for non-empty name with
valid characters.

This separation allows to be specific with errors in number of cases.

An existing function to validate keyspace name is shared between
several classes, and a unit test of it is added.